2022 day24 and 25
