--- advent_of_code/2022/README.md
@@ -1,6 +1,0 @@
-Build command:
-`wasm-pack build --scope nickymeuleman --target web --out-dir www/pkg`
-
-to publish to npm:
-`cd www/pkg`
-`npm publish --access=public`
