--- advent_of_code/aoc/2022/inputs/day24_sample.txt
@@ -1,6 +1,0 @@
-#.######
-#>>.<^<#
-#.<..<<#
-#>v.><>#
-#<^v^^>#
-######.#
