--- advent_of_code/aoc/2022/inputs/day25_sample.txt
@@ -1,13 +1,0 @@
-1=-0-2
-12111
-2=0=
-21
-2=01
-111
-20012
-112
-1=-1=
-1-12
-12
-1=
-122
